--- src/@types/blog.d.ts
@@ -1,20 +1,0 @@
-export interface BlogPostTeaser {
-  id: string;
-  frontmatter: {
-    title: string;
-    featuredImage: any;
-    date: string;
-    path: string;
-    teaserText: string;
-  };
-}
-
-export interface LocalesQuery {
-  edges: {
-    node: {
-      ns: string;
-      language: string;
-      data: string;
-    };
-  }[];
-}
