+++ src/pages/blog.tsx
@@ -13,7 +13,7 @@ import { getImage, IGatsbyImageData } from 'gatsby-plugin-image';
 import { useI18next, useTranslation } from 'gatsby-plugin-react-i18next';
 import { ButtonText, SendButton } from '../components/form/controls';
 import { RightArrowIcon } from '../components/icons/buttons-icons/right-arrow';
-import { LocalesQuery } from '../@types/blog';
+import { LocalesQuery } from '../types';
 
 const TOP_POST_COUNT = 2;
