+++ src/templates/career-details.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import SEO from '../components/seo';
-import { SyPersonioJob } from '../@types/personio';
 import { graphql } from 'gatsby';
 import { useTranslation } from 'gatsby-plugin-react-i18next';
 import { CareerDetails } from '../page-building/career-details/career-details';
+import { SyPersonioJob } from '../types';
 
 interface CareerPageProps {
   pageContext: {
