+++ src/types.ts
@@ -1,3 +1,24 @@
+export interface BlogPostTeaser {
+  id: string;
+  frontmatter: {
+    title: string;
+    featuredImage: any;
+    date: string;
+    path: string;
+    teaserText: string;
+  };
+}
+
+export interface LocalesQuery {
+  edges: {
+    node: {
+      ns: string;
+      language: string;
+      data: string;
+    };
+  }[];
+}
+
 /**
  * the normalized jobs content coming from the API
  */
